Challenge_5/cloud_run_functions/agent-communication/requirements.txt
functions-framework==3.*
pandas
google-cloud-bigquery
google-cloud-aiplatform
flask
db-dtypes
